requirements.txt
azure-functions==1.21.3
requests>=2.31.0
psycopg2-binary==2.9.9
